factor change observation into `observe-changes` module; logical data storage into `logical-tree` module; remove cruft; start on detecting when alternate patching scheme is needed.
